separate preSave handlers
